requirements.txt
streamlit 
langchain
langchain_groq
langchain_community
SQLAchemy
python_dotenv
tabulate